fix: revoke model access cache when model created
